Move window metrics helpers to WindowSize.kt
