+++ JetNews/app/src/main/java/com/example/jetnews/ui/JetnewsApp.kt
@@ -16,9 +16,6 @@
 
 package com.example.jetnews.ui
 
-import android.app.Activity
-import android.content.Context
-import android.content.ContextWrapper
 import androidx.compose.material.DrawerState
 import androidx.compose.material.DrawerValue
 import androidx.compose.material.MaterialTheme
@@ -26,24 +23,16 @@ import androidx.compose.material.ModalDrawer
 import androidx.compose.material.rememberDrawerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.SideEffect
-import androidx.compose.runtime.State
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalDensity
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import androidx.window.layout.WindowInfoRepository
-import androidx.window.layout.WindowInfoRepository.Companion.windowInfoRepository
-import androidx.window.layout.WindowMetrics
-import androidx.window.layout.WindowMetricsCalculator
 import com.example.jetnews.data.AppContainer
 import com.example.jetnews.ui.theme.JetnewsTheme
 import com.example.jetnews.utils.WindowSize
-import com.example.jetnews.utils.getWindowSize
+import com.example.jetnews.utils.rememberWindowSizeState
 import com.google.accompanist.insets.ProvideWindowInsets
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import kotlinx.coroutines.launch
@@ -100,48 +89,6 @@ fun JetnewsApp(
         }
     }
 }
-
-/**
- * Returns the [WindowMetrics] corresponding to the current activities
- * [WindowInfoRepository.currentWindowMetrics] as [State].
- */
-@Composable
-private fun rememberCurrentWindowMetrics(): State<WindowMetrics> {
-    val activity = LocalContext.current.findActivity()
-    val currentWindowMetricsFlow = remember(activity) {
-        activity.windowInfoRepository().currentWindowMetrics
-    }
-    return currentWindowMetricsFlow.collectAsState(
-        WindowMetricsCalculator.getOrCreate().computeCurrentWindowMetrics(activity)
-    )
-}
-
-/**
- * Remembers the [WindowSize] corresponding to the current window metrics.
- */
-@Composable
-private fun rememberWindowSizeState(): WindowSize {
-    val windowMetrics by rememberCurrentWindowMetrics()
-    val density = LocalDensity.current
-
-    return remember(windowMetrics, density) {
-        with(density) {
-            getWindowSize(windowMetrics.bounds.width().toDp())
-        }
-    }
-}
-
-/**
- * Find the closest Activity in a given Context.
- */
-private tailrec fun Context.findActivity(): Activity =
-    when (this) {
-        is Activity -> this
-        is ContextWrapper -> baseContext.findActivity()
-        else -> throw IllegalStateException(
-            "findActivity should be called in the context of an Activity"
-        )
-    }
 
 /**
  * Determine the drawer state to pass to the modal drawer.
